section x and y to float

+++ db/migrate/20210216220344_add_x_to_sections.rb
@@ -1,5 +1,5 @@
 class AddXToSections < ActiveRecord::Migration[6.1]
   def change
-    add_column :sections, :x, :integer
+    add_column :sections, :x, :float
   end
 end
